Organizing methods in DuckDuckGoSmartForm and SmartFormService
